Refactor logic for initializing multistage
In order to ensure that we correctly create the stage files only when
necessary and only set the default stage when a stagefile exists, this
commit changes what happens when the plugin is loaded. It moves the
majority of the logic into simple methods with descriptive names so that
the intent is much more clear. Additionally, it fixes a bug where
initialize would bomb out because it was trying to set the default stage
with a file that didn't exist yet. And finally, the last change it makes
is to remove setting the rails environment in each stage. Because this
project is designed to be used with more than just rails applications,
adding this by default wasn't a really sane default.