Refactor default stages
In order to make the stages created simpler, this commit changes how we
decide what the default stages are. In addition, it adds logging during
the stage creation process and ensures it is idempotent so that we don't
overwrite existing files. And finally, it updates the readme to include
more information.